LayoutTests/media/modern-media-controls/airplay-support/airplay-support-expected.txt

Testing the AirPlaySupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Making AirPlay routes available.
PASS mediaController.controls.airplayButton.enabled is true
PASS mediaController.controls.airplayButton.on is false

Entering AirPlay playback mode.
PASS mediaController.controls.airplayButton.enabled is true
PASS mediaController.controls.airplayButton.on is true

PASS successfullyParsed is true

TEST COMPLETE
